+++ examples/resources/cockroach_metric_export_prometheus_config/resource.tf
@@ -1,0 +1,7 @@
+variable "cluster_id" {
+  type = string
+}
+
+resource "cockroach_metric_export_prometheus_config" "example" {
+  id = var.cluster_id
+}